taking out annoying comments

+++ hw5/msort.h
@@ -67,7 +67,7 @@ void multMergeSort(std::vector<T>& myArray, int k, int low, int high, Comparator
         // }
 
         std::cout << "Indicies: ";
-        for (int i = 0; i < index.size(); i++) std::cout << index[i] << " ";
+        for (size_t i = 0; i < index.size(); i++) std::cout << index[i] << " ";
         	std::cout << std::endl;
 
         for (size_t i = 0; i < index.size() - 1; i++) {
@@ -77,42 +77,7 @@ void multMergeSort(std::vector<T>& myArray, int k, int low, int high, Comparator
         merge(myArray, index, m, comp);
     }
 }
-    // if (low <= high) {
-    //     int rem = 0;
-    //     //if ((high - low) % k != 0) rem = (high- - low) % k;
-        
-    //     int m = (high-low + 1)/k;
 
-    //     std::vector<int> index;
-
-    //     for (size_t i = 0; i < k; i++) {
-    //         index.push_back(m*i);
-    //     }
-
-
-    //     // int cnt = rem;
-    //     // for (size_t j = 0; j < index.size(); j++) {
-    //     //     if (cnt != 0 && index[j] != 0) {
-    //     //     	index[j] -= 1;
-    //     //     	cnt--;
-    //     //     }
-    //     // }
-    //     for (size_t i = 0; i < index.size(); i++) std::cout << index[i] << " ";
- 	// 		std::cout << std::endl;
-
-    //     for(int i = 0; i < index.size() - 1; i++){
-    //     	std::cout << "current start is " << index[i] << "and current end is " << index[i+1] -1 << std::endl;
-	// 		multMergeSort(myArray, k, index[i], index[i+1]-1, comp);
-    //     }
-
-     	//multMergeSort(myArray,k,index[index.size()-1],high,comp);
- 		
- 		//return;''
-        // multMergeSort(myArray, k, index[0], index[1], comp);
-        // multMergeSort(myArray, k, index[1]+1, index[2], comp);
-        //merge(myArray, index, m, comp);
-//     }
-// }
 
 template <class T, class Comparator>
 void mergeSort (std::vector<T>& myArray, int k, Comparator comp) {
